requirements.txt
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.16
django-cors-headers==3.5.0
django-environ==0.4.5
django-resized==0.3.11
django-rest-swagger==2.2.0
djangorestframework==3.12.1
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient==2.0.1
openapi-codec==1.3.2
Pillow==8.0.0
pkg-resources==0.0.0
PyMySQL==0.10.1
pytz==2020.1
requests==2.24.0
simplejson==3.17.2
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.25.11